Revert "added chat history helper"
This reverts commit c9b27ab8fbb8ec6b4538242668f55f2b2f4753ee.
