--- Explorer/Assets/DCL/Chat/ChatControllerMessageHandlingHelper.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d3d57bc82a6bbf14fb584ed4fb8335c0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
